Change track capture rules to match playback more closely.
The set of tracks now matches exactly what is present in the source.
Tracks that don't have any content are reflected using the "muted" attribute.
This allows for capture of media that has a track that is selected, but not
currently playing.  It also cleanly deals with seeking because the set of tracks
doesn't need to change when seeking, just the muted property.

We had no good way to handle disabled and muted tracks previously.  This
corrects that.

captureStreamUntilEnded() just waits until playback ends.  It isn't affected
by source changes or seeking.

Also switched canvas capture to capture nothing when not origin-clean.